+++ src/apps/molresponse/TDDFT.cc
@@ -3656,481 +3656,6 @@ std::vector<real_function_3d> TDDFT::GetConjugateTransitionDensities(
   return densities;
 }
 
-template <std::size_t NDIM>
-void TDDFT::set_protocol(World& world, double thresh) {
-  size_t k;
-  // Allow for imprecise conversion of threshold
-  if (thresh >= 0.9e-2)
-    k = 4;
-  else if (thresh >= 0.9e-4)
-    k = 6;
-  else if (thresh >= 0.9e-6)
-    k = 8;
-  else if (thresh >= 0.9e-8)
-    k = 10;
-  else
-    k = 12;
-
-  // k defaults to make sense with thresh, override by providing k in
-  // input file
-  if (r_params.k() > 0) {
-    FunctionDefaults<NDIM>::set_k(r_params.k());
-  } else {
-    FunctionDefaults<NDIM>::set_k(k);
-  }
-
-  // MolDFT sets all these, so copying
-  FunctionDefaults<NDIM>::set_thresh(thresh);
-  FunctionDefaults<NDIM>::set_refine(true);
-  FunctionDefaults<NDIM>::set_initial_level(2);
-  FunctionDefaults<NDIM>::set_autorefine(false);
-  FunctionDefaults<NDIM>::set_apply_randomize(false);
-  FunctionDefaults<NDIM>::set_project_randomize(false);
-  // GaussianConvolution1DCache<double>::map.clear();//(TODO:molresponse-What
-  // is this? Do i need it?)
-
-  // dconv defaults to thresh*100, overrirde by providing dconv in input
-  // file
-  if (r_params.dconv_set() == false) {
-    r_params.set_derived_value<double>("dconv", thresh * 100);
-  }
-
-  // Basic print
-  if (world.rank() == 0) {
-    print("\nSolving NDIM=",
-          NDIM,
-          " with thresh",
-          thresh,
-          "    k",
-          FunctionDefaults<NDIM>::get_k(),
-          "  dconv",
-          std::max(thresh, r_params.dconv()),
-          "\n");
-  }
-}
-
-void TDDFT::check_k(World& world, double thresh, size_t k) {
-  // Boolean to redo ground hamiltonian calculation if
-  // ground state orbitals change
-  bool redo = false;
-
-  // Verify ground state orbitals have correct k
-  if (FunctionDefaults<3>::get_k() != ground_orbitals[0].k()) {
-    // Re-read orbitals from the archive (assuming
-    // the archive has orbitals stored at a higher
-    // k value than what was previously computed
-    // with)
-    g_params.read(world, r_params.archive());
-    reconstruct(world, ground_orbitals);
-
-    // Reset correct k (its set in g_params.read)
-    FunctionDefaults<3>::set_k(k);
-
-    // Project each ground state to correct k
-    for (unsigned int i = 0; i < ground_orbitals.size(); i++)
-      ground_orbitals[i] = project(
-          ground_orbitals[i], FunctionDefaults<3>::get_k(), thresh, false);
-    world.gop.fence();
-
-    // Clean up a bit
-    truncate(world, ground_orbitals);
-
-    // Ground state orbitals changed, clear old hamiltonian
-    redo = true;
-  }
-
-  // Recalculate ground state hamiltonian here
-  if (redo or !hamiltonian.has_data()) {
-    hamiltonian =
-        CreateGroundHamiltonian(world, ground_orbitals, r_params.print_level());
-  }
-
-  // If we stored the potential, check that too
-  if (r_params.store_potential()) {
-    if (FunctionDefaults<3>::get_k() != stored_potential[0][0].k()) {
-      // Project the potential into correct k
-      for (unsigned int i = 0; i < stored_potential.size(); i++) {
-        reconstruct(world, stored_potential[i]);
-        for (unsigned int j = 0; j < stored_potential[0].size(); j++)
-          stored_potential[i][j] = project(stored_potential[i][j],
-                                           FunctionDefaults<3>::get_k(),
-                                           thresh,
-                                           false);
-        world.gop.fence();
-      }
-    }
-    if (FunctionDefaults<3>::get_k() != stored_v_coul.k())
-      stored_v_coul =
-          project(stored_v_coul, FunctionDefaults<3>::get_k(), thresh, false);
-    if (FunctionDefaults<3>::get_k() != stored_v_nuc.k())
-      stored_v_nuc =
-          project(stored_v_nuc, FunctionDefaults<3>::get_k(), thresh, false);
-  }
-
-  // Verify response functions have correct k
-  if (Chi.X.size() != 0) {
-    if (FunctionDefaults<3>::get_k() != Chi.X[0][0].k()) {
-      // Project all x components into correct k
-      for (unsigned int i = 0; i < Chi.X.size(); i++) {
-        reconstruct(world, Chi.X[i]);
-        for (unsigned int j = 0; j < Chi.X[0].size(); j++)
-          Chi.X[i][j] =
-              project(Chi.X[i][j], FunctionDefaults<3>::get_k(), thresh, false);
-        world.gop.fence();
-      }
-      Chi.X.truncate_rf();
-
-      // Do same for y components if applicable
-      // (Always do this, as y will be zero
-      //  and still used in doing DFT and TDA)
-      // Project all y components into correct k
-      for (unsigned int i = 0; i < Chi.Y.size(); i++) {
-        reconstruct(world, Chi.Y[i]);
-        for (unsigned int j = 0; j < Chi.Y[0].size(); j++)
-          Chi.Y[i][j] =
-              project(Chi.Y[i][j], FunctionDefaults<3>::get_k(), thresh, false);
-        world.gop.fence();
-      }
-      Chi.Y.truncate_rf();
-    }
-  }
-  // Verify response functions have correct k
-  if (Chi.X.size() != 0) {
-    if (FunctionDefaults<3>::get_k() != Chi.X[0][0].k()) {
-      // Project all x components into correct k
-      for (unsigned int i = 0; i < Chi.X.size(); i++) {
-        reconstruct(world, Chi.X[i]);
-        for (unsigned int j = 0; j < Chi.X[0].size(); j++)
-          Chi.X[i][j] =
-              project(Chi.X[i][j], FunctionDefaults<3>::get_k(), thresh, false);
-        world.gop.fence();
-      }
-      Chi.X.truncate_rf();
-
-      // Do same for y components if applicable
-      // (Always do this, as y will be zero
-      //  and still used in doing DFT and TDA)
-      // Project all y components into correct k
-      for (unsigned int i = 0; i < Chi.Y.size(); i++) {
-        reconstruct(world, Chi.Y[i]);
-        for (unsigned int j = 0; j < Chi.Y[0].size(); j++)
-          Chi.Y[i][j] =
-              project(Chi.Y[i][j], FunctionDefaults<3>::get_k(), thresh, false);
-        world.gop.fence();
-      }
-      Chi.Y.truncate_rf();
-    }
-  }
-
-  // Don't forget the mask function as well
-  if (FunctionDefaults<3>::get_k() != mask.k()) {
-    mask = project(mask, FunctionDefaults<3>::get_k(), thresh, false);
-  }
-  // Don't forget right hand side
-
-  // Verify response functions have correct k
-  if (PQ.X.size() != 0) {
-    if (FunctionDefaults<3>::get_k() != PQ.X[0][0].k()) {
-      // Project all x components into correct k
-      for (unsigned int i = 0; i < Chi.X.size(); i++) {
-        reconstruct(world, PQ.X[i]);
-        for (unsigned int j = 0; j < PQ.X[0].size(); j++)
-          PQ.X[i][j] =
-              project(PQ.X[i][j], FunctionDefaults<3>::get_k(), thresh, false);
-        world.gop.fence();
-      }
-      Chi.X.truncate_rf();
-
-      // Do same for y components if applicable
-      // (Always do this, as y will be zero
-      //  and still used in doing DFT and TDA)
-      // Project all y components into correct k
-      for (unsigned int i = 0; i < PQ.Y.size(); i++) {
-        reconstruct(world, PQ.Y[i]);
-        for (unsigned int j = 0; j < PQ.Y[0].size(); j++)
-          PQ.Y[i][j] =
-              project(PQ.Y[i][j], FunctionDefaults<3>::get_k(), thresh, false);
-        world.gop.fence();
-      }
-      Chi.Y.truncate_rf();
-    }
-  }
-  // Make sure everything is done before leaving
-  world.gop.fence();
-}
-
-// Creates random guess functions semi-intelligently(?)
-response_space TDDFT::create_random_guess(
-    World& world,
-    size_t m,                                // m response states
-    size_t n,                                // n ground states
-    std::vector<real_function_3d>& grounds,  // guess should have size n
-    Molecule& molecule) {
-  // Basic output
-  if (world.rank() == 0)
-    print("   Using a random guess for initial response functions.\n");
-
-  // Create empty container and add in randomness
-  response_space f(world, m, n);
-  f = add_randomness(world, f, 1e3);  // noise all over the world
-
-  // Create and apply a centered gaussian on each atom so that the
-  // randomness is localized around the atoms
-  real_function_3d gaus = real_factory_3d(world);
-  for (auto atom : molecule.get_atoms()) {
-    real_function_3d x =
-        real_factory_3d(world).functor(real_functor_3d(new GaussianGuess<3>(
-            atom.get_coords(), 0.01, std::vector<int>{0, 0, 0})));
-    gaus = gaus + x;
-  }
-  f = f * gaus;
-
-  // Project out groundstate from guesses
-  QProjector<double, 3> projector(world, grounds);
-  for (unsigned int i = 0; i < f.size(); i++) f[i] = projector(f[i]);
-
-  // Normalize
-  normalize(world, f);
-
-  return f;
-}
-
-// Creates random guess functions semi-intelligently(?)
-std::vector<real_function_3d> TDDFT::create_random_guess(
-    World& world,
-    size_t m,
-    std::vector<real_function_3d>& grounds,
-    Molecule& molecule) {
-  // Basic output
-  if (world.rank() == 0)
-    print("   Using a random guess for initial response functions.");
-
-  // Create empty container and add in randomness
-  std::vector<real_function_3d> f =
-      zero_functions_compressed<double, 3>(world, m);
-  // create vector of m functions:w
-
-  // Create and apply a centered gaussian on each atom so that the
-  // randomness is localized around the atoms
-  real_function_3d gaus = real_factory_3d(world);
-  for (auto atom : molecule.get_atoms()) {
-    real_function_3d x =
-        real_factory_3d(world).functor(real_functor_3d(new GaussianGuess<3>(
-            atom.get_coords(), 0.01, std::vector<int>{0, 0, 0})));
-    gaus = gaus + x;
-  }
-
-  // Lambda function to add in noise
-  auto lambda = [](const Key<3>& key, Tensor<double>& x) mutable {
-    Tensor<double> y(x.size());
-    y.fillrandom();
-    y.scale(1e3);
-    x = x + y;
-  };
-
-  // Go through each function in f_copy and add in random noise
-  for (unsigned int i = 0; i < f.size(); i++) {
-    // Add in random noise using rng and a the defined lambda function
-    f[i].unaryop(lambda);
-
-    // Apply mask to get boundary condition right
-    f[i] = mask * f[i] * gaus;
-  }
-
-  // Project out groundstate from guesses
-  QProjector<double, 3> projector(world, grounds);
-  for (unsigned int i = 0; i < f.size(); i++) f[i] = projector(f[i]);
-
-  // Normalize
-  for (unsigned int i = 0; i < f.size(); i++) {
-    double norm = f[i].norm2();
-    f[i].scale(1.0 / norm);
-  }
-
-  return f;
-}
-
-// Creates an initial guess function from nwchem output files
-response_space TDDFT::create_nwchem_guess(World& world, size_t m) {
-  // Basic output
-  if (world.rank() == 0)
-    print("   Creating an initial guess from NWChem file", r_params.nwchem());
-
-  // Create empty containers
-  response_space f;
-
-  // Create the nwchem reader
-  slymer::NWChem_Interface nwchem(r_params.nwchem(), std::cout);
-
-  // For parallel runs, silencing all but 1 slymer instance
-  if (world.rank() != 0) {
-    std::ostream dev_null(nullptr);
-    nwchem.err = dev_null;
-  }
-
-  // Read in basis set
-  nwchem.read(slymer::Properties::Basis);
-
-  // Read in the molecular orbital coefficients, energies,
-  // and occupancies
-  nwchem.read(slymer::Properties::MOs | slymer::Properties::Energies |
-              slymer::Properties::Occupancies);
-
-  // Create the nwchem orbitals as madness functions
-  std::vector<real_function_3d> temp1;
-  for (auto basis :
-       slymer::cast_basis<slymer::GaussianFunction>(nwchem.basis_set)) {
-    // Get the center of gaussian as its special point
-    std::vector<coord_3d> centers;
-    coord_3d r;
-    r[0] = basis.get().center[0];
-    r[1] = basis.get().center[1];
-    r[2] = basis.get().center[2];
-    centers.push_back(r);
-
-    // Now make the function
-    temp1.push_back(FunctionFactory<double, 3>(world).functor(
-        std::shared_ptr<FunctionFunctorInterface<double, 3>>(
-            new slymer::Gaussian_Functor(basis.get(), centers))));
-
-    // Let user know something is going on
-    if (temp1.size() % 10 == 0 and world.rank() == 0)
-      print("Created", temp1.size(), "functions.");
-  }
-  if (world.rank() == 0) print("Finished creating", temp1.size(), "functions.");
-
-  // Normalize ao's
-  madness::normalize(world, temp1);
-
-  // Transform ao's now
-  std::vector<real_function_3d> temp = madness::transform(
-      world, temp1, nwchem.MOs, FunctionDefaults<3>::get_thresh(), true);
-
-  // Now save the unoccupied orbitals
-  std::vector<real_function_3d> temp2;
-  size_t num_virt = 0;
-  for (size_t i = 0; i < temp1.size(); i++) {
-    if (nwchem.occupancies[i] == 0) {
-      temp2.push_back(temp[i]);
-      num_virt++;
-    }
-  }
-
-  // Create as many vectors of functions as we can from these nwchem
-  // virtual orbitals putting 1 virtual orbital from nwchem per vector
-  for (size_t i = 0; i < std::max(m, num_virt); i++) {
-    // Create the vector to add the new function to
-    std::vector<real_function_3d> v1 =
-        zero_functions_compressed<double, 3>(world, g_params.orbitals().size());
-
-    // Put the "new" function into the vector
-    v1[i % v1.size()] = temp2[i];
-
-    // Add vector to return container
-    f.push_back(v1);
-
-    // See if we've made enough functions
-    if (f.size() >= m) break;
-  }
-  if (world.rank() == 0)
-    print("Created", f.size(), "guess functions from provided NWChem data.");
-
-  // If not enough functions have been made, start adding symmetry adapted
-  // functions
-  size_t n = f.size();
-
-  // If still not enough functions have been made, add in random guesses
-  if (n < m) {
-    // Tell user the bad news
-    if (world.rank() == 0)
-      print("\n   Only",
-            n,
-            "guess functions were provided by augmenting NWChem "
-            "functions.\n   "
-            "Augmenting with random functions.");
-
-    // Create the random guess
-    Molecule mol = g_params.molecule();
-    response_space rand =
-        create_random_guess(world, m - n, n, ground_orbitals, mol);
-
-    // Add to vector of functions
-    for (unsigned int i = 0; i < rand.size(); i++) f.push_back(rand[i]);
-  }
-
-  // Project out groundstate from guesses
-  QProjector<double, 3> projector(world, ground_orbitals);
-  for (unsigned int i = 0; i < f.size(); i++) f[i] = projector(f[i]);
-
-  // Truncate and normalize
-  f.truncate_rf();
-  normalize(world, f);
-
-  return f;
-}
-
-// Creates potentials using the ResponsePotential object
-// Potentials are modified in place
-void TDDFT::create_all_potentials(World& world,
-                                  response_space& x,
-                                  response_space& x_gamma,
-                                  response_space& x_V0,
-                                  ResponsePotential& potentials,
-                                  size_t print_level) {
-  // Intermediaries
-  response_space gammaJ, gammaK, groundJ, groundK;
-
-  // Calc. coulomb like terms
-  if (print_level >= 1) molresponse::start_timer(world);
-  potentials.coulomb_terms(x, gammaK, groundJ);
-  if (print_level >= 1) molresponse::end_timer(world, "Coulomb terms:");
-
-  // Calc. exchange like terms
-  if (print_level >= 1) molresponse::start_timer(world);
-  potentials.exchange_terms(x, gammaJ, groundK);
-  if (print_level >= 1) molresponse::end_timer(world, "Exchange terms:");
-
-  // Assemble pieces together
-  x_gamma = gammaJ - gammaK;
-  x_V0 = groundJ - groundK;
-
-  // Debugging output
-  if (print_level >= 2) {
-    // Coulomb
-    if (world.rank() == 0) printf("   Coulomb Deriv matrix:\n");
-    Tensor<double> temp = expectation(world, x, gammaJ);
-    if (world.rank() == 0) print(temp);
-
-    // Exchange or VXC
-    if (r_params.xc() == "hf" and world.rank() == 0)
-      printf("   Exchange Deriv matrix:\n");
-    if (r_params.xc() != "hf" and world.rank() == 0)
-      printf("   Negative of XC Deriv matrix:\n");
-    temp = expectation(world, x, gammaK);
-    if (world.rank() == 0) print(temp);
-
-    // Total Gamma
-    if (world.rank() == 0) printf("   Gamma matrix:\n");
-    temp = expectation(world, x, x_gamma);
-    if (world.rank() == 0) print(temp);
-
-    // Coulomb (ground)
-    if (world.rank() == 0) printf("   Coulomb + Nuclear potential matrix:\n");
-    temp = expectation(world, x, groundJ);
-    if (world.rank() == 0) print(temp);
-
-    // Exchange or VXC (ground)
-    if (r_params.xc() == "hf" and world.rank() == 0)
-      printf("   Exchange potential matrix:\n");
-    if (r_params.xc() != "hf" and world.rank() == 0)
-      printf("   XC potential matrix:\n");
-    temp = expectation(world, x, groundK);
-    if (world.rank() == 0) print(temp);
-    if (world.rank() == 0) printf("   Total Potential Energy matrix:\n");
-    temp = expectation(world, x, x_V0);
-    if (world.rank() == 0) print(temp);
-  }
-}
 
 void TDDFT::polarizability(World& world, Tensor<double> polar) {
   // Get transition density
@@ -4327,125 +3852,6 @@ void TDDFT::plot_excited_states(World& world,
 // No matter the calculation type we do the same iteration.
 // The only difference is the number of response states as well as the
 // number of right hand side vectors.
-void TDDFT::compute_freq_response(World& world) {
-  molresponse::start_timer(world);
-  std::string property = r_params.response_type();
-  // Warm and fuzzy
-  if (world.rank() == 0) {
-    print("\n\n    Response Calculation");
-    print("   ------------------------");
-  }
-
-  // Keep a copy of dipoles * MO (needed explicitly in eq.)
-
-  // For each protocol
-  for (unsigned int proto = 0; proto < r_params.protocol().size(); proto++) {
-    // Set defaults inside here
-    // default value of
-    set_protocol<3>(world, r_params.protocol()[proto]);
-
-    check_k(world, r_params.protocol()[proto], FunctionDefaults<3>::get_k());
-    // Do something to ensure all functions have same k value
-
-    if (property.compare("dipole") == 0) {
-      if (world.rank() == 0) print("creating dipole property operator");
-      p = Property(world, "dipole");
-    } else if (property.compare("nuclear") == 0) {
-      if (world.rank() == 0) print("creating nuclear property operator");
-      p = Property(world, "nuclear", g_params.molecule());
-    }
-
-    // Create guesses if no response functions
-    // If restarting, load here
-    if (proto == 0) {
-      if (r_params.restart()) {
-        if (world.rank() == 0)
-          print("   Initial guess from file:", r_params.restart_file());
-        load(world, r_params.restart_file());
-        check_k(
-            world, r_params.protocol()[proto], FunctionDefaults<3>::get_k());
-
-        if (r_params.dipole()) {
-          // set states
-          PQ.X = PropertyRHS(world, p);
-          PQ.Y = PQ.X.copy();
-          print("P: ", PQ.X.norm2());
-          print("Q: ", PQ.Y.norm2());
-
-          // set RHS_Vector
-        } else if (r_params.nuclear()) {
-          // set guesses
-          // print("Creating X for Nuclear Operators");
-
-          // print("x norms:");
-          // print(x_response.norm2());
-          PQ.X = PropertyRHS(world, p);
-          PQ.Y = PQ.X.copy();
-
-        } else if (r_params.order2()) {
-          //
-        } else if (r_params.order3()) {
-          //
-        } else {
-          MADNESS_EXCEPTION("not a valid response state ", 0);
-        }
-      } else {  // Dipole guesses
-
-        if (r_params.dipole()) {
-          PQ.X = PropertyRHS(world, p);
-          PQ.Y = PQ.X.copy();
-          // set states
-          //
-          print("okay this is not a good idea if it comes up more than once");
-          // set RHS_Vector
-        } else if (r_params.nuclear()) {
-          PQ.X = PropertyRHS(world, p);
-          PQ.Y = PQ.X.copy();
-        } else if (r_params.order2()) {
-          //
-        } else if (r_params.order3()) {
-          //
-        } else {
-          MADNESS_EXCEPTION("not a valid response state ", 0);
-        }
-      }
-    }
-    //
-    // Here i should print some information about the calculation we are
-    // about to do
-    print("Preiteration Information");
-    print("Property : ", property);
-    print("Number of Response States: ", r_params.n_states());
-    print("Number of Ground States: ", r_params.num_orbitals());
-    print("k = ", FunctionDefaults<3>::get_k());
-    print("protocol threshold = ", FunctionDefaults<3>::get_k());
-
-    print("Property rhs func k = ", PQ.X[0][0].k());
-    print("Property func k thresh= ", PQ.X[0][0].thresh());
-
-    print("Property rhs func Q k = ", PQ.Y[0][0].k());
-    print("Property func Q k thresh = ", PQ.Y[0][0].thresh());
-
-    print("Property rhs func P norms", PQ.X.norm2());
-    print("Property rhs func Q norms", PQ.Y.norm2());
-
-    // Now actually ready to iterate...
-    iterate_freq(world);
-    // IterateFrequencyResponse(world, P, Q);
-  }  // end for --finished reponse density
-
-  // Have response function, now calculate polarizability for this axis
-  // polarizability(world, polar_tensor);
-  // PrintPolarizabilityAnalysis(world, polar_tensor, omega);
-
-  // Print total time
-  // Precision is set to 10 coming in, drop it to 2
-  std::cout.precision(2);
-  std::cout << std::fixed;
-
-  // Get start time
-  molresponse::end_timer(world, "total:");
-}  // end compute frequency response
 
 // Exactam eam
 // Deuces
